+++ init.go
@@ -53,6 +53,7 @@ func init() {
 		panic(err)
 	}
 
+	// Because ffi_get_version and ffi_get_version_number just exist since libffi 3.5.0, we don't panic here.
 	getVersion, _ = libffi.Get("ffi_get_version")
 
 	getVersionNumber, _ = libffi.Get("ffi_get_version_number")