+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- libffi 3.5.0 now lets you retrieve the version at runtime. Therefore the following new functions have been added:
+    - `func GetVersion() string`
+    - `func GetVersionNumber() uint64`
+
 ## [0.4.1] - 2025-05-12
 
 ### Fixed
@@ -59,6 +66,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Dependencies `github.com/ebitengine/purego` and `golang.org/x/sys` updated.
 
+[Unreleased]: https://github.com/JupiterRider/ffi/compare/v0.4.1...main
 [0.4.1]: https://github.com/JupiterRider/ffi/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/JupiterRider/ffi/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/JupiterRider/ffi/compare/v0.2.2...v0.3.0